cloudprober: catch tls handshake errors (#13)
* cloudprober: catch tls handshake errors

Catch tls handshake errors by adding two more
conditions to the isSSLError. The first is adding
a check to see if there is a record header error.
The next is to check if the error string has a prefix
of "tls: " as this denotes a handshake error in the
go standard library.

* cloudprober: add extra checks for isSSLError

add a check to make sure the error is not nil before
proceeding.